Update on Database

+++ src/de/fhl/overchef/model/Picture.java
@@ -1,11 +1,16 @@
 package de.fhl.overchef.model;
 
+import java.awt.image.BufferedImage;
 import java.io.*;
 import java.nio.file.Files;
 import java.nio.file.Paths;
 
 import javax.imageio.ImageIO;
+import javax.imageio.stream.ImageOutputStream;
 
+import de.fhl.overchef.view.RecipeModifyController;
+import javafx.embed.swing.SwingFXUtils;
+import javafx.scene.control.Alert;
 import javafx.scene.image.Image;
 /**
  * This Class is used to save and process image, such as loading, storing and deleting image.
@@ -15,11 +20,20 @@ import javafx.scene.image.Image;
 public class Picture {
 	int width;
 	int length;
+	String res = "//src//de//fhl//overchef//model//Pictures//";
+	String pictureName = "default name";
 	String root = "default";
 	Image image=null;
+	BufferedImage bufferImage = null;
 	FileOutputStream imageOut;
 	File file;
 	
+	public Image getImage() {
+		return image;
+	}
+	public void setImage(Image image) {
+		this.image = image;
+	}
 	public String getRoot() {
 		return root;
 	}
@@ -33,7 +47,9 @@ public class Picture {
 	public Picture(String path){
 		try{
 			 image = new Image(new FileInputStream(path));		
+//			 System.out.println("Picture " + image.getHeight());
 			 file = new File(path);
+			 this.pictureName = file.getName();
 		}
 		catch(Exception e){}
 	}
@@ -43,11 +59,13 @@ public class Picture {
 	 * @throws FileNotFoundException
 	 * @throws IOException
 	 */
-	public boolean checkName() throws FileNotFoundException, IOException {
-		File[] fileList = new File(System.getProperty("user.dir") + "\\src\\de\\fhl\\overchef\\model\\Pictures\\").listFiles();
+	public static boolean checkName(String path) throws FileNotFoundException, IOException {
+		//create a file linked to the file chosen by user, if user does't choose a file, parameter path will be default value.
+		File chosenFile = new File(path);
+		File[] fileList = new File(System.getProperty("user.dir") + "//src//de//fhl//overchef//model//Pictures//").listFiles();
 		for(File f:fileList) {
 			System.out.println(f.getName());
-			if(f.getName().equals(file.getName())) {		
+			if(f.getName().equals(chosenFile.getName())) {		
 				System.out.println("Same image name in current directory");
 				return false;
 			}
@@ -60,16 +78,33 @@ public class Picture {
 	 * @throws IOException
 	 */
 	public void writePicture() throws IOException {
-		imageOut = new FileOutputStream(this.root = System.getProperty("user.dir") + "\\src\\de\\fhl\\overchef\\model\\Pictures\\" + file.getName());
-		imageOut.write(Files.readAllBytes(Paths.get(this.root)));	
-		imageOut.close();
+		System.out.println(file.getName());
+		System.out.println("Picture " + root);
+
+			File outputImage = new File(this.root = System.getProperty("user.dir") + res + file.getName());
+			bufferImage = SwingFXUtils.fromFXImage(image, null);
+			imageOut  = new FileOutputStream(outputImage);
+			ImageIO.write(bufferImage, pictureName.substring(pictureName.lastIndexOf(".") + 1), imageOut);
+			imageOut.close();
 	}
 	/**
 	 * delete the picture
 	 */
 	public void deletePicture() {
 		this.file = new File(this.root);
-		this.file.delete();
+		if (file.exists() && !file.getName().equals("OverChefDefaultPicture.jpg")) {
+			this.file.delete();
+			// when the user delete the picture, root should be change to a default path, otherwise program will have an exception after save the recipe
+			if (!this.file.exists()) {
+				this.root = System.getProperty("user.dir") + res + "OverChefDefaultPicture.jpg";
+			}
+		}else {
+			Alert noFileAlert = new Alert(Alert.AlertType.INFORMATION);
+			noFileAlert.setTitle("Reminder");
+			noFileAlert.setHeaderText("No file to delete");
+			noFileAlert.showAndWait();
+			System.out.println("From Picture: Deleting a file: No file to delete");
+		}
 	}
 	
 	public void deletePicture(String path) {
